Graded HW5
Docstrings are present for functions, but are not detailed enough or describe self-commenting code. Be sure to explain where numbers are coming if they don't obviously come from earlier in the code. Nice use of in line comments at points, but there are instances of unnecessary comments. **Documentation: +1**

There are a few unnecessary classes and functions that could be condensed. Define Classes in the order they are used to allow for a logical flow of your code. **Style: +0.9**

Code runs as intended. Game is super cute and fun to play, though I couldn't seem to get any adult dinosaurs and the people/dino speed need to be adjusted for balanced gameplay. **Functionality: +3**

Nicely done! :)
Grade: 4.9/5